+++ tencentcloud/common/http/request.go
@@ -254,7 +254,7 @@ func CompleteCommonParams(request Request, region string) {
 	params["Action"] = request.GetAction()
 	params["Timestamp"] = strconv.FormatInt(time.Now().Unix(), 10)
 	params["Nonce"] = strconv.Itoa(rand.Int())
-	params["RequestClient"] = "SDK_GO_1.0.412"
+	params["RequestClient"] = "SDK_GO_1.0.413"
 }
 
 func ConstructParams(req Request) (err error) {