+++ tencentcloud/lighthouse/v20200324/errors.go
@@ -452,6 +452,9 @@ const (
 	// 快照忙。
 	UNSUPPORTEDOPERATION_SNAPSHOTBUSY = "UnsupportedOperation.SnapshotBusy"
 
+	// 系统忙。
+	UNSUPPORTEDOPERATION_SYSTEMBUSY = "UnsupportedOperation.SystemBusy"
+
 	// Windows实例不支持绑定密钥对。
 	UNSUPPORTEDOPERATION_WINDOWSNOTALLOWTOASSOCIATEKEYPAIR = "UnsupportedOperation.WindowsNotAllowToAssociateKeyPair"
 )