+++ tencentcloud/nlp/v20190408/errors.go
@@ -26,6 +26,9 @@ const (
 	// 后端服务超时。
 	FAILEDOPERATION_REQUESTTIMEOUT = "FailedOperation.RequestTimeout"
 
+	// RPC请求失败，一般为算法微服务故障。
+	FAILEDOPERATION_RPCFAIL = "FailedOperation.RpcFail"
+
 	// 文本向量化失败
 	FAILEDOPERATION_TEXTEMBEDDINGFAILED = "FailedOperation.TextEmbeddingFailed"