+++ tencentcloud/tiia/v20190529/models.go
@@ -99,6 +99,24 @@ func (r *AssessQualityResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }
 
+type Attribute struct {
+
+	// 属性
+	Type *string `json:"Type,omitempty" name:"Type"`
+
+	// 属性详情
+	Details *string `json:"Details,omitempty" name:"Details"`
+}
+
+type Box struct {
+
+	// 图像主体区域。
+	Rect *ImageRect `json:"Rect,omitempty" name:"Rect"`
+
+	// 置信度。
+	Score *float64 `json:"Score,omitempty" name:"Score"`
+}
+
 type CarPlateContent struct {
 
 	// 车牌信息。
@@ -146,6 +164,18 @@ type CarTagItem struct {
 	PlateContent *CarPlateContent `json:"PlateContent,omitempty" name:"PlateContent"`
 }
 
+type ColorInfo struct {
+
+	// RGB颜色值（16进制），例如：291A18。
+	Color *string `json:"Color,omitempty" name:"Color"`
+
+	// 当前颜色标签所占比例。
+	Percentage *float64 `json:"Percentage,omitempty" name:"Percentage"`
+
+	// 颜色标签。蜜柚色，米驼色等。
+	Label *string `json:"Label,omitempty" name:"Label"`
+}
+
 type Coord struct {
 
 	// 横坐标x
@@ -173,7 +203,7 @@ type CreateGroupRequest struct {
 	// 访问限制默认为10qps，如需扩容请联系[在线客服](https://cloud.tencent.com/online-service)申请。
 	MaxQps *uint64 `json:"MaxQps,omitempty" name:"MaxQps"`
 
-	// 图库类型，对应不同服务类型，默认为1。建议手动调整为4～6，1～3为历史版本，不推荐。
+	// 图库类型，对应不同服务类型，默认为4。1～3为历史版本，不推荐。
 	// 参数值：
 	// 4：在自建图库中搜索相同原图，可支持裁剪、翻转、调色、加水印后的图片搜索，适用于图片版权保护、原图查询等场景。
 	// 5：在自建图库中搜索相同或相似的商品图片，适用于商品分类、检索、推荐等电商场景。
@@ -241,22 +271,46 @@ type CreateImageRequest struct {
 
 	// 图片的 Url 。对应图片 base64 编码后大小不可超过5M。  
 	// Url、Image必须提供一个，如果都提供，只使用 Url。 
-	// 图片分辨率不超过4096*4096。
+	// 图片分辨率不超过4096\*4096。
 	// 图片存储于腾讯云的Url可保障更高下载速度和稳定性，建议图片存储于腾讯云。 
 	// 非腾讯云存储的Url速度和稳定性可能受一定影响。 
 	// 支持PNG、JPG、JPEG、BMP，不支持 GIF 图片。
+	// 注意：开启主体识别分辨率不超过2000\*2000，图片长宽比小于10（长/短 < 10）。
 	ImageUrl *string `json:"ImageUrl,omitempty" name:"ImageUrl"`
-
-	// 图片 base64 数据，base64 编码后大小不可超过5M。 
-	// 图片分辨率不超过4096*4096。 
-	// 支持PNG、JPG、JPEG、BMP，不支持 GIF 图片。
-	ImageBase64 *string `json:"ImageBase64,omitempty" name:"ImageBase64"`
 
 	// 用户自定义的内容，最多支持4096个字符，查询时原样带回。
 	CustomContent *string `json:"CustomContent,omitempty" name:"CustomContent"`
 
+	// 图片 base64 数据，base64 编码后大小不可超过5M。 
+	// 图片分辨率不超过4096\*4096。 
+	// 支持PNG、JPG、JPEG、BMP，不支持 GIF 图片。
+	// 注意：开启主体识别分辨率不超过2000\*2000，图片长宽比小于10（长/短 < 10）。
+	ImageBase64 *string `json:"ImageBase64,omitempty" name:"ImageBase64"`
+
 	// 图片自定义标签，最多不超过10个，格式为JSON。
 	Tags *string `json:"Tags,omitempty" name:"Tags"`
+
+	// 是否需要启用主体识别，默认为**TRUE**。
+	// 1.  为**TRUE**时，启用主体识别，返回主体信息。若没有指定**ImageRect**，自动提取最大面积主体创建图片并进行主体识别。主体识别结果可在**Response**中获取。
+	// 2. 为**FALSE**时，不启用主体识别，不返回主体信息。若没有指定**ImageRect**，以整张图创建图片。
+	// 注意：服务类型为商品图像搜索时生效。
+	EnableDetect *bool `json:"EnableDetect,omitempty" name:"EnableDetect"`
+
+	// 图像类目ID。
+	// 若设置类目ID，提取对应类目的主体创建图片。
+	// 注意：服务类型为商品图像搜索时生效。
+	// 类目信息：
+	// 0：上衣。
+	// 1：裙装。
+	// 2：下装。
+	// 3：包。
+	// 4：鞋。
+	// 5：配饰。
+	CategoryId *int64 `json:"CategoryId,omitempty" name:"CategoryId"`
+
+	// 图像主体区域。
+	// 若设置主体区域，提取指定的区域创建图片。
+	ImageRect *Rect `json:"ImageRect,omitempty" name:"ImageRect"`
 }
 
 func (r *CreateImageRequest) ToJsonString() string {
@@ -275,9 +329,12 @@ func (r *CreateImageRequest) FromJsonString(s string) error {
 	delete(f, "EntityId")
 	delete(f, "PicName")
 	delete(f, "ImageUrl")
-	delete(f, "ImageBase64")
 	delete(f, "CustomContent")
+	delete(f, "ImageBase64")
 	delete(f, "Tags")
+	delete(f, "EnableDetect")
+	delete(f, "CategoryId")
+	delete(f, "ImageRect")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "CreateImageRequest has unknown keys!", "")
 	}
@@ -287,6 +344,12 @@ func (r *CreateImageRequest) FromJsonString(s string) error {
 type CreateImageResponse struct {
 	*tchttp.BaseResponse
 	Response *struct {
+
+		// 输入图的主体信息。
+	// 若启用主体识别且在请求中指定了类目ID或主体区域，以指定的主体为准。若启用主体识别且没有指定，以最大面积主体为准。
+	// 注意：此字段可能返回 null，表示取不到有效值。服务类型为商品图像搜索时生效。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+		Object *ObjectInfo `json:"Object,omitempty" name:"Object"`
 
 		// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
 		RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
@@ -1301,6 +1364,21 @@ type Location struct {
 	YMax *int64 `json:"YMax,omitempty" name:"YMax"`
 }
 
+type ObjectInfo struct {
+
+	// 图像主体区域。
+	Box *Box `json:"Box,omitempty" name:"Box"`
+
+	// 主体类别ID。
+	CategoryId *int64 `json:"CategoryId,omitempty" name:"CategoryId"`
+
+	// 整张图颜色信息。
+	Colors []*ColorInfo `json:"Colors,omitempty" name:"Colors"`
+
+	// 属性信息。
+	Attributes []*Attribute `json:"Attributes,omitempty" name:"Attributes"`
+}
+
 type Product struct {
 
 	// 图片中商品的三级分类识别结果，选取所有三级分类中的置信度最大者
@@ -1491,6 +1569,21 @@ func (r *RecognizeCarResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }
 
+type Rect struct {
+
+	// x轴坐标
+	X *int64 `json:"X,omitempty" name:"X"`
+
+	// y轴坐标
+	Y *int64 `json:"Y,omitempty" name:"Y"`
+
+	// (x,y)坐标距离长度
+	Width *int64 `json:"Width,omitempty" name:"Width"`
+
+	// (x,y)坐标距离高度
+	Height *int64 `json:"Height,omitempty" name:"Height"`
+}
+
 type RegionDetected struct {
 
 	// 商品的品类预测结果。 
@@ -1510,33 +1603,54 @@ type SearchImageRequest struct {
 	// 图库名称。
 	GroupId *string `json:"GroupId,omitempty" name:"GroupId"`
 
-	// 图片的 Url 。对应图片 base64 编码后大小不可超过2M。 
-	// 图片分辨率不超过1920*1080。 
+	// 图片的 Url 。对应图片 base64 编码后大小不可超过5M。 
+	// 图片分辨率不超4096\*4096。 
 	// Url、Image必须提供一个，如果都提供，只使用 Url。 
 	// 图片存储于腾讯云的Url可保障更高下载速度和稳定性，建议图片存储于腾讯云。 
 	// 非腾讯云存储的Url速度和稳定性可能受一定影响。 
 	// 支持PNG、JPG、JPEG、BMP，不支持 GIF 图片。
+	// 注意：开启主体识别分辨率不超过2000\*2000，图片长宽比小于10（长/短 < 10）。
 	ImageUrl *string `json:"ImageUrl,omitempty" name:"ImageUrl"`
 
-	// 图片 base64 数据，base64 编码后大小不可超过2M。 
-	// 图片分辨率不超过1920*1080。 
+	// 图片 base64 数据，base64 编码后大小不可超过5M。 
+	// 图片分辨率不超过4096\*4096。 
 	// 支持PNG、JPG、JPEG、BMP，不支持 GIF 图片。
+	// 注意：开启主体识别分辨率不超过2000\*2000，图片长宽比小于10（长/短 < 10）。
 	ImageBase64 *string `json:"ImageBase64,omitempty" name:"ImageBase64"`
 
-	// 出参Score中，只有超过MatchThreshold值的结果才会返回。默认为0
-	MatchThreshold *int64 `json:"MatchThreshold,omitempty" name:"MatchThreshold"`
+	// 返回数量，默认值为10，最大值为100。
+	Limit *int64 `json:"Limit,omitempty" name:"Limit"`
 
 	// 起始序号，默认值为0。
 	Offset *int64 `json:"Offset,omitempty" name:"Offset"`
 
-	// 返回数量，默认值为10，最大值为100。
-	Limit *int64 `json:"Limit,omitempty" name:"Limit"`
+	// 出参Score中，只有超过**MatchThreshold**值的结果才会返回。默认为0
+	MatchThreshold *int64 `json:"MatchThreshold,omitempty" name:"MatchThreshold"`
 
 	// 针对入库时提交的Tags信息进行条件过滤。支持>、>=、 <、 <=、=，!=，多个条件之间支持AND和OR进行连接。
 	Filter *string `json:"Filter,omitempty" name:"Filter"`
 
 	// 图像主体区域。
+	// 若设置主体区域，提取指定的区域进行检索。
 	ImageRect *ImageRect `json:"ImageRect,omitempty" name:"ImageRect"`
+
+	// 是否需要启用主体识别，默认为**TRUE** 。
+	// 1. 为**TRUE**时，启用主体识别，返回主体信息。若没有指定**ImageRect**，自动提取最大面积主体进行检索并进行主体识别。主体识别结果可在**Response中**获取。
+	// 2. 为**FALSE**时，不启用主体识别，不返回主体信息。若没有指定**ImageRect**，以整张图检索图片。
+	// 注意：服务类型为商品图像搜索时生效。
+	EnableDetect *bool `json:"EnableDetect,omitempty" name:"EnableDetect"`
+
+	// 图像类目ID。
+	// 若设置类目ID，提取对应类目的主体进行检索。
+	// 注意：服务类型为商品图像搜索时生效。
+	// 类目信息：
+	// 0：上衣。
+	// 1：裙装。
+	// 2：下装。
+	// 3：包。
+	// 4：鞋。
+	// 5：配饰。
+	CategoryId *int64 `json:"CategoryId,omitempty" name:"CategoryId"`
 }
 
 func (r *SearchImageRequest) ToJsonString() string {
@@ -1554,11 +1668,13 @@ func (r *SearchImageRequest) FromJsonString(s string) error {
 	delete(f, "GroupId")
 	delete(f, "ImageUrl")
 	delete(f, "ImageBase64")
-	delete(f, "MatchThreshold")
-	delete(f, "Offset")
 	delete(f, "Limit")
+	delete(f, "Offset")
+	delete(f, "MatchThreshold")
 	delete(f, "Filter")
 	delete(f, "ImageRect")
+	delete(f, "EnableDetect")
+	delete(f, "CategoryId")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "SearchImageRequest has unknown keys!", "")
 	}
@@ -1575,6 +1691,12 @@ type SearchImageResponse struct {
 		// 图片信息。
 	// 注意：此字段可能返回 null，表示取不到有效值。
 		ImageInfos []*ImageInfo `json:"ImageInfos,omitempty" name:"ImageInfos"`
+
+		// 输入图的主体信息。
+	// 若启用主体识别且在请求中指定了类目ID或主体区域，以指定的主体为准。若启用主体识别且没有指定，以最大面积主体为准。
+	// 注意：此字段可能返回 null，表示取不到有效值。服务类型为商品图像搜索时生效。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+		Object *ObjectInfo `json:"Object,omitempty" name:"Object"`
 
 		// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
 		RequestId *string `json:"RequestId,omitempty" name:"RequestId"`